修复了timer_interrupt的bug（只要qemu保证能在0.1s内完成stvec的写入）

+++ include/lib/print.h
@@ -9,5 +9,5 @@ void printf(const char* fmt, ...);
 // void assert(bool condition, const char* warning);
 
 void panic(const char *fmt, ...);
-void assert(int condition, const char* warning, ...);
+void assert(uint64 condition, const char* warning, ...);
 #endif
